agents/loopie/requirements.txt
google-adk>=1.27.0
asyncpg>=0.29.0
